MVP of Calculator, Flickd, and Cat Finder
Calculator current working. The answer will show during a chaining
of operations or if the equal button is pushed in between. Numbers
start at the left and the front number does not bleed. Negative/
positive button adds or removes a negative sign, however, it needs
to be clicked twice in order to remove a negative sign from
a negative number that was displayed as a result.

Flickd has a black on white display, with a title and subtitle that
appear at the top of the screen, and scale/wrap upon resizing. The
search bar has placeholder text and is appened to the top of the screen
so results occure underneath. Upon searching, the new items appear at the
top of the images class.

Click on a bush in cat finder, and either a pumpkin or a cat is revealed.
The image fades to none as the other image fades to all. Once the cat
image is found, a dialogue on the script blinks, asking the user to
click on it which refreshes the page